refactor crypto
