+++ accumulators/src/merkle_tree.rs
@@ -3,7 +3,7 @@ use noah_algebra::{
     collections::{hash_map::Iter, HashMap},
     prelude::*,
 };
-use noah_crypto::basic::anemoi_jive::{AnemoiJive, AnemoiJive381, ANEMOI_JIVE_381_SALTS};
+use noah_crypto::anemoi_jive::{AnemoiJive, AnemoiJive381, ANEMOI_JIVE_381_SALTS};
 use storage::db::MerkleDB;
 use storage::store::{ImmutablePrefixedStore, PrefixedStore, Stated, Store};
