+++ api/benches/merkle_tree.rs
@@ -7,7 +7,7 @@ use noah::anon_xfr::structs::{AccElemVars, MTLeafInfo, MTNode, MTPath, MerklePat
 use noah_accumulators::merkle_tree::{PersistentMerkleTree, Proof, TreePath};
 use noah_algebra::bls12_381::BLSScalar;
 use noah_algebra::prelude::*;
-use noah_crypto::basic::anemoi_jive::{
+use noah_crypto::anemoi_jive::{
     AnemoiJive, AnemoiJive381, AnemoiVLHTrace, JiveTrace, ANEMOI_JIVE_381_SALTS,
 };
 use noah_plonk::plonk::constraint_system::{TurboCS, VarIndex};
