+++ api/src/anon_xfr/abar_to_abar.rs
@@ -27,9 +27,7 @@ use crate::parameters::{
 use digest::{consts::U64, Digest};
 use merlin::Transcript;
 use noah_algebra::{bls12_381::BLSScalar, prelude::*};
-use noah_crypto::basic::anemoi_jive::{
-    AnemoiJive, AnemoiJive381, AnemoiVLHTrace, ANEMOI_JIVE_381_SALTS,
-};
+use noah_crypto::anemoi_jive::{AnemoiJive, AnemoiJive381, AnemoiVLHTrace, ANEMOI_JIVE_381_SALTS};
 use noah_plonk::plonk::{
     constraint_system::{TurboCS, VarIndex},
     prover::prover_with_lagrange,
@@ -1114,7 +1112,7 @@ mod tests {
     use digest::Digest;
     use merlin::Transcript;
     use noah_algebra::{bls12_381::BLSScalar, prelude::*};
-    use noah_crypto::basic::anemoi_jive::{
+    use noah_crypto::anemoi_jive::{
         AnemoiJive, AnemoiJive381, AnemoiVLHTrace, ANEMOI_JIVE_381_SALTS,
     };
     use noah_plonk::plonk::constraint_system::{TurboCS, VarIndex};
