+++ api/src/xfr/proofs.rs
@@ -22,7 +22,8 @@ use noah_algebra::{
     utils::{min_greater_equal_power_of_two, u64_to_u32_pair},
 };
 use noah_crypto::{
-    basic::{
+    bulletproofs::range::{batch_verify_ranges, prove_ranges},
+    {
         chaum_pedersen::{
             chaum_pedersen_batch_verify_multiple_eq, chaum_pedersen_prove_multiple_eq,
             ChaumPedersenProofX,
@@ -33,7 +34,6 @@ use noah_crypto::{
             PedersenElGamalEqProof, PedersenElGamalProofInstance,
         },
     },
-    bulletproofs::range::{batch_verify_ranges, prove_ranges},
 };
 
 const POW_2_32: u64 = 0xFFFF_FFFFu64 + 1;
