--- crypto/src/basic/mod.rs
@@ -1,14 +1,0 @@
-/// The module for the Anemoi-Jive hash functions.
-pub mod anemoi_jive;
-/// The module for the Chaum-Pedersen protocol.
-pub mod chaum_pedersen;
-/// The module for the doubly SNARK-friendly primitives.
-pub mod doubly_snark_friendly;
-/// The module for the ElGamal encryption.
-pub mod elgamal;
-/// The module for hybrid encryption.
-pub mod hybrid_encryption;
-/// The module for the matrix Sigma protocol.
-pub mod matrix_sigma;
-/// The module for the equality proof between a Pedersen commitment and an ElGamal ciphertext.
-pub mod pedersen_elgamal;
